超过4GB的文件，字节溢出的问题。
Former-commit-id: 88b972f5e1025300b2c81e27c67b56f55e62b0f2

+++ ReadySHAREDemo/KxSMBProvider.h
@@ -83,7 +83,7 @@ typedef void (^KxSMBBlockProgress)(KxSMBItem *item, unsigned long transferred);
 @interface KxSMBItemStat : NSObject
 @property(readonly, nonatomic, strong) NSDate *lastModified;
 @property(readonly, nonatomic, strong) NSDate *lastAccess;
-@property(readonly, nonatomic) unsigned long size;
+@property(readonly, nonatomic) unsigned long long size;
 @property(readonly, nonatomic) long mode;
 @end
